Balloon Suck and Walker Death
Implemented sucking in with balloon and consuming walkers.
Walkers now damage the player when hit as well.

+++ MYBL_AB/Arglib.h
@@ -135,7 +135,7 @@ struct Rect
     int x;
     int y;
     uint8_t width;
-    uint8_t height;
+    int height;
 };
 
 struct Point
